custom: nfd: add pull policy flag
in order to avoid to skip uploading images on quay
(cumbersome: require quay account, takes time and bandwidth)
we need to change the default pull policy.

Leverage the flag to do that, in a backward compatible way:
the script will work with unpatched NFDs, and in this case
the e2e test will just fail (!).

Signed-off-by: Francesco Romani <fromani@redhat.com>